Move Chart class to new file
