+++ test/utils/chart_state_utils.dart
@@ -1,5 +1,5 @@
 import 'package:flutter_test/flutter_test.dart';
-import 'package:time_chart/time_chart.dart';
+import 'package:time_chart/src/chart.dart';
 
 ChartState getChartState(WidgetTester tester) {
   return tester.state(find.byType(Chart));
